cleaned up unused files and added detailed comments for an easier read during later review
