--- src/Components/Score.js
@@ -1,5 +1,0 @@
-
-
-const addToScoreButton = () => {
-    
-}
